Changed method signature to take long instead of int for size

+++ pdisk-server/war/src/main/java/eu/stratuslab/storage/disk/plugins/DiskStorage.java
@@ -2,9 +2,9 @@ package eu.stratuslab.storage.disk.plugins;
 
 public interface DiskStorage {
 
-    public void create(String uuid, int size);
+    public void create(String uuid, long size);
 
-    public void createCopyOnWrite(String baseUuid, String cowUuid, int size);
+    public void createCopyOnWrite(String baseUuid, String cowUuid, long size);
 
     public String rebase(String uuid, String rebaseUuid);
