+++ pdisk-server/war/src/main/java/eu/stratuslab/storage/disk/utils/FileUtils.java
@@ -78,7 +78,7 @@ public final class FileUtils {
 		}
 	}
 
-	public static void createZeroFile(File file, int sizeInGB) {
+	public static void createZeroFile(File file, long sizeInGB) {
 		OutputStream ostream = null;
 
 		try {
